+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+
+- The devtools console now confirms when CSS files have been successfully loaded after a previous error
+
 ## [0.27.0] - 2023-06-01
 
 ### Fixed